+++ mc-queens-dealership-be/likes.py
@@ -1,15 +1,17 @@
 from flask import jsonify, Blueprint, request
 from models import Like as ModelLike, db
 from sqlalchemy import exc
+from flask_login import login_required, current_user
 
 likes_routes = Blueprint("likes", __name__)
 
 @likes_routes.route("/like/create", methods=["POST"])
+@login_required
 def like_create():
     try:
         like = ModelLike(
-            user_id = request.form["user_id"],
-            car_id = request.form["car_id"],
+            user_id = current_user.id,
+            car_id = request.form.get("car_id"),
         )
         db.session.add(like)
         db.session.commit()
@@ -19,8 +21,11 @@ def like_create():
 
 
 @likes_routes.route("/like/<int:id>/delete", methods=["DELETE"])
+@login_required
 def like_delete(id):
     like = ModelLike.query.get_or_404(id)
+    if not like.user_id == current_user.id:
+        return jsonify({"err": ";P"})
     try:
         db.session.delete(like)
         db.session.commit()
